pyproject.toml
[project]
name = "community-cloud-storage"
version = "0.4.0"
description = "Distributed storage for community archives using ipfs-cluster, Tailscale and Docker"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.0",
]
packages = ["community_cloud_storage"]

[project.urls]
repository = "https://github.com/historypin/community-cloud-storage"

[project.scripts]
community-cloud-storage = "community_cloud_storage:cli"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"
